Generative-AI/sarvam/requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
annoy==1.17.3
anyio==4.5.0
attrs==24.2.0
beautifulsoup4==4.12.3
bitarray==2.9.2
bitsandbytes==0.42.0
blinker==1.8.2
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colbert-ai==0.2.19
dataclasses-json==0.6.7
datasets==3.0.0
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
et-xmlfile==1.1.0
faiss-cpu==1.8.0.post1
fast-pytorch-kmeans==0.2.0.1
fastapi==0.115.0
filelock==3.16.1
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.6.1
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.1
langchain-text-splitters==0.3.0
langsmith==0.1.123
llama-cloud==0.0.17
llama-index==0.11.10
llama-index-agent-openai==0.3.3
llama-index-cli==0.3.1
llama-index-core==0.11.10
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.3.1
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.9
llama-index-multi-modal-llms-openai==0.2.1
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.6
llvmlite==0.43.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
onnx==1.16.2
openai==1.46.1
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pacmap==0.7.3
pandas==2.2.2
pillow==10.4.0
protobuf==5.28.2
psutil==6.0.0
pyarrow==17.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pynvml==11.5.3
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
RAGatouille==0.0.8.post4
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==2.7.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
srsly==2.4.8
starlette==0.38.5
striprtf==0.0.26
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.5
transformers==4.44.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.30.6
voyager==2.0.9
Werkzeug==3.0.4
wrapt==1.16.0
xxhash==3.5.0
yarl==1.11.1
